Module Github.Gist

The Gist module provides access to the GitHub gist API.

val for_user : ?token:Token.t -> ?since:string -> user:string -> unit -> Github_t.gist Stream.t

for_user ?since ~user () is a stream of gists that belong to user. If ?since is an ISO 8601 timestamp, only gists updated since this time are returned.

val all : ?token:Token.t -> ?since:string -> unit -> Github_t.gist Stream.t

all ?since () is a stream of all of the gists for the current token's user or all public gists if invoked without a current token. If ?since is an ISO 8601 timestamp, only gists updated since this time are returned.

val all_public : ?token:Token.t -> ?since:string -> unit -> Github_t.gist Stream.t

all_public ?since () is a stream of all of the public gists for the current token's user or all public gists if invoked without a current token. If ?since is an ISO 8601 timestamp, only gists updated since this time are returned.

val starred : ?token:Token.t -> ?since:string -> unit -> Github_t.gist Stream.t

starred ?since () is a stream of all starred gists for the current token's user. If ?since is an ISO 8601 timestamp, only gists updated since this time are returned.

val get : ?token:Token.t -> id:string -> unit -> Github_t.gist Response.t Monad.t

get ~id () is the gist id.

val create : ?token:Token.t -> gist:Github_t.new_gist -> unit -> Github_t.gist Response.t Monad.t

create ~gist () is a newly created gist described by gist.

val update : ?token:Token.t -> id:string -> gist:Github_t.update_gist -> unit -> Github_t.gist Response.t Monad.t

update ~id ~gist () is the updated gist id as described by gist.

val commits : ?token:Token.t -> id:string -> unit -> Github_t.gist_commit Stream.t

commits ~id () is a stream of commits for gist id.

val star : ?token:Token.t -> id:string -> unit -> unit Response.t Monad.t

star ~id () activates after gist id is marked as starred by the current token's user.

val unstar : ?token:Token.t -> id:string -> unit -> unit Response.t Monad.t

unstar ~id () activates after gist id is marked as not starred by the current token's user.

val fork : ?token:Token.t -> id:string -> unit -> Github_t.gist Response.t Monad.t

fork ~id () is a newly forked gist from gist id.

val forks : ?token:Token.t -> id:string -> unit -> Github_t.gist_fork Stream.t

forks ~id () is a stream of forks of gist id.

val delete : ?token:Token.t -> id:string -> unit -> unit Response.t Monad.t

delete ~id () activates after gist id has been deleted.