Module Github.Organization

The Organization module exposes the functionality of the GitHub organization API.

val teams : ?⁠token:Token.t -> org:string -> unit -> Github_t.team Stream.t

teams ~org () is a stream of teams belonging to the organization org.

val user_orgs : ?⁠token:Token.t -> user:string -> unit -> Github_t.org Stream.t

user_orgs ~user () is a stream of the organizations to which the user user belongs.

val current_user_orgs : ?⁠token:Token.t -> unit -> Github_t.org Stream.t

current_user () is a stream of the organizations to which the user linked to current token belongs, and for which the user granted access to the organizations to the current token.

val repositories : ?⁠token:Token.t -> org:string -> unit -> Github_t.repository Stream.t

repositories ~org () is a stream of repositories belonging to the organization org.

module Hook : sig ... end

The Hook module provides access to GitHub's organization webhooks API which lets you manage an organization's remote notification hooks.