Module Github.Stats

The Stats module exposes the functionality of GitHub's repository statistics API which provides historical data regarding the aggregate behavior of a repository.

val contributors : ?⁠token:Token.t -> user:string -> repo:string -> unit -> Github_t.contributor_stats Stream.t

contributors ~user ~repo () is a stream of all contributor statistics for user/repo. The stream is empty if the data are not cached yet